requirements.txt
Flask==0.10.1
Flask-Bootstrap==3.2.0.2
Flask-Cache==0.13.1
Jinja2==2.7.3
MarkupSafe==0.23
PyYAML==3.11
SPARQLWrapper==1.6.4
Werkzeug==0.9.6
arrow==0.4.4
future==0.14.1
gevent==1.0.1
gevent-socketio==0.3.6
gevent-websocket==0.9.3
greenlet==0.4.5
gunicorn==19.1.1
html5lib==0.999
isodate==0.5.0
itsdangerous==0.24
memoized-property==1.0.2
mendeley==0.2.2
oauthlib==0.6.3
pyparsing==2.0.3
python-dateutil==2.2
rdflib==4.1.2
requests==2.4.3
requests-oauthlib==0.4.2
six==1.8.0
wsgiref==0.1.2
